% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_sce}
\alias{example_sce}
\title{Load Example sce dataset}
\usage{
example_sce(
  n_row = 10,
  n_col = 10,
  n_gene = 1000,
  lambda = 2,
  frac_sparse = 0.95
)
}
\arguments{
\item{n_row}{integer; number of simulated spatial pixel rows}

\item{n_col}{integer; number of simulated spatial pixel columns}

\item{n_gene}{integer; number of simulated genes}

\item{lambda}{numeric; poisson parameter}

\item{frac_sparse}{numeric; fraction of 0s.}
}
\value{
SummarizedExperiment object
}
\description{
Load Example sce dataset
}
