% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_clust.R
\name{smooth_clust}
\alias{smooth_clust}
\title{Cluster the right singular vectors from model output}
\usage{
smooth_clust(v, n_clust)
}
\arguments{
\item{v}{matrix; right singular vectors}

\item{n_clust}{integer; number of clusters.}
}
\value{
vector of cluster labels
}
\description{
Cluster the right singular vectors from model output
}
\examples{

sce <- example_sce()
x <- SummarizedExperiment::assay(sce)
coords <- SummarizedExperiment::colData(sce)[, c("col", "row")]
k <- 10
init <- smooth_init(x, coords, k)
lambda <- 1
epsilon <- 1e-3
maxiter <- 5
model <- smooth_model(x, init$u0, init$v0, init$w, init$index, lambda, epsilon, maxiter)
n_clust <- 3
smooth_clust(model$v, n_clust)

}
