% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_lrc.R
\name{smooth_lrc}
\alias{smooth_lrc}
\title{smoothLRC}
\usage{
smooth_lrc(input, lambda, k, n_clust, epsilon = 0.001, maxiter = 1000)
}
\arguments{
\item{input}{SummarizedExperiment object containing counts assay and row/col coordinates.}

\item{lambda}{positive numeric; penalization parameter.}

\item{k}{integer; indicates desired rank of singular value decomposition.}

\item{n_clust}{integer; number of clusters.}

\item{epsilon}{positive numeric; convergence criterion.}

\item{maxiter}{positive integer; maximum desired iterations}
}
\value{
SummarizedExperiment object with u, v and cluster labels.
}
\description{
smoothLRC
}
\examples{

sce <- example_sce()
lambda <- 5
k <- 10
n_clust <- 5
epsilon <- 1e-3
maxiter <- 5
smooth_lrc(sce, lambda, k, n_clust, epsilon, maxiter)

}
