% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clust.R
\name{spatial_clust}
\alias{spatial_clust}
\title{Main Spatial Clustering Method}
\usage{
spatial_clust(
  input,
  u_init,
  v_init,
  w = NULL,
  index = NULL,
  lambda = 1,
  cnorm = NULL,
  k = 20,
  epsilon = 0.001,
  maxiter = 1000
)
}
\arguments{
\item{input}{SingleCellExperiment object containing counts assay and row/col coordinates.}

\item{u_init}{matrix; u initialization matrix}

\item{v_init}{matrix; v initialization matrix}

\item{w}{matrix; distance matrix. If null, computed to specification with k nearest neighbors.}

\item{index}{w index.}

\item{lambda}{numeric; penalization parameter. set to NULL for lambda selection.}

\item{cnorm}{a numeric value for the amount of V normalization. NULL for no normalization.}

\item{k}{integer; if knn = TRUE, number of nearest neighbors to consider}

\item{epsilon}{numeric; convergence criterion}

\item{maxiter}{integer; maximum number of iterations}
}
\value{
Object containing u, v and convergence information
}
\description{
Main Spatial Clustering Method
}
