% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_clust.R
\name{smooth_clust}
\alias{smooth_clust}
\title{Clustering of Right Singular Vectors}
\usage{
smooth_clust(v, n_clust)
}
\arguments{
\item{v}{matrix; right singular vectors}

\item{n_clust}{integer; number of clusters.}
}
\value{
vector of cluster labels
}
\description{
Clustering of Right Singular Vectors
}
