% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cv.R
\name{smooth_cv}
\alias{smooth_cv}
\title{Parameter Selection via Cross-Validation}
\usage{
smooth_cv(input, lambda, k, test_size, epsilon = 0.001, maxiter = 1000)
}
\arguments{
\item{input}{SummarizedExperiment object containing counts assay and row/col coordinates.}

\item{lambda}{positive numeric; penalization parameter.}

\item{k}{integer; indicates desired rank of singular value decomposition.}

\item{test_size}{The number of pixels to use in the test set.}

\item{epsilon}{positive numeric; convergence criterion.}

\item{maxiter}{positive integer; maximum desired iterations}
}
\value{
A list containing the penalized likelihood from the model and the cross-validated likelihood from the test set.
}
\description{
Parameter Selection via Cross-Validation
}
\examples{
sce <- example_sce()
smooth_cv(sce, 1, 10, 100, maxiter = 10)
}
