% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_init.R
\name{smooth_init}
\alias{smooth_init}
\title{Initialization for smoothLRC}
\usage{
smooth_init(x, coords, k, knn = 6)
}
\arguments{
\item{x}{matrix; count assay.}

\item{coords}{matrix; spatial coordinates.}

\item{k}{integer; indicates desired rank of singular value decomposition.}

\item{knn}{integer; nearest neighbors.}
}
\value{
list of objects for smooth_model function.
}
\description{
Initialization for smoothLRC
}
\examples{

sce <- example_sce()
x <- SummarizedExperiment::assay(sce)
coords <- SummarizedExperiment::colData(sce)[, c("col", "row")]
k <- 10
smooth_init(x, coords, k)

}
